+++ ui/src/admin/reducers/influxdb.js
@@ -1,6 +1,5 @@
 import reject from 'lodash/reject'
 import {
-  NEW_DEFAULT_USER,
   NEW_DEFAULT_ROLE,
   NEW_DEFAULT_DATABASE,
   NEW_EMPTY_RP,
@@ -67,14 +66,6 @@ const adminInfluxDB = (state = initialState, action) => {
       return {...state, ...action.payload}
     }
 
-    case 'INFLUXDB_ADD_USER': {
-      const newUser = {...NEW_DEFAULT_USER, isEditing: true}
-      return {
-        ...state,
-        users: [newUser, ...state.users],
-      }
-    }
-
     case 'INFLUXDB_ADD_ROLE': {
       const newRole = {...NEW_DEFAULT_ROLE, isEditing: true}
       return {
@@ -115,11 +106,13 @@ const adminInfluxDB = (state = initialState, action) => {
 
     case 'INFLUXDB_SYNC_USER': {
       const {staleUser, syncedUser} = action.payload
-      const newState = {
-        users: state.users.map(u =>
-          u.links.self === staleUser.links.self ? {...syncedUser} : u
-        ),
-      }
+      const newState = staleUser.links
+        ? {
+            users: state.users.map(u =>
+              u.links.self === staleUser.links.self ? {...syncedUser} : u
+            ),
+          }
+        : {users: [{...syncedUser}, ...state.users]}
       return {...state, ...newState}
     }
 
@@ -159,16 +152,6 @@ const adminInfluxDB = (state = initialState, action) => {
         ),
       }
 
-      return {...state, ...newState}
-    }
-
-    case 'INFLUXDB_EDIT_USER': {
-      const {user, updates} = action.payload
-      const newState = {
-        users: state.users.map(u =>
-          u.links.self === user.links.self ? {...u, ...updates} : u
-        ),
-      }
       return {...state, ...newState}
     }
 
@@ -237,11 +220,13 @@ const adminInfluxDB = (state = initialState, action) => {
 
     case 'INFLUXDB_DELETE_USER': {
       const {user} = action.payload
-      const newState = {
-        users: state.users.filter(u => u.links.self !== user.links.self),
+      if (user.links) {
+        const newState = {
+          users: state.users.filter(u => u.links.self !== user.links.self),
+        }
+        return {...state, ...newState}
       }
-
-      return {...state, ...newState}
+      return state
     }
 
     case 'INFLUXDB_DELETE_ROLE': {